Move Checkers back in the VersionUpdater namespace
