+++ src/Configuration.php
@@ -19,6 +19,9 @@ use Psy\Output\ShellOutput;
 use Psy\TabCompletion\AutoCompleter;
 use Psy\VarDumper\Presenter;
 use Psy\VersionUpdater\Checker;
+use Psy\VersionUpdater\GithubChecker;
+use Psy\VersionUpdater\IntervalChecker;
+use Psy\VersionUpdater\NoopChecker;
 use Symfony\Component\Console\Formatter\OutputFormatterStyle;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -1562,7 +1565,7 @@ class Configuration
             $interval = $this->getUpdateCheck();
             switch ($interval) {
                 case Checker::ALWAYS:
-                    $this->checker = new Checker\GitHubChecker();
+                    $this->checker = new GitHubChecker();
                     break;
 
                 case Checker::DAILY:
@@ -1570,14 +1573,14 @@ class Configuration
                 case Checker::MONTHLY:
                     $checkFile = $this->getUpdateCheckCacheFile();
                     if ($checkFile === false) {
-                        $this->checker = new Checker\NoopChecker();
+                        $this->checker = new NoopChecker();
                     } else {
-                        $this->checker = new Checker\IntervalChecker($checkFile, $interval);
+                        $this->checker = new IntervalChecker($checkFile, $interval);
                     }
                     break;
 
                 case Checker::NEVER:
-                    $this->checker = new Checker\NoopChecker();
+                    $this->checker = new NoopChecker();
                     break;
             }
         }
