+++ src/VersionUpdater/GitHubChecker.php
@@ -9,10 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Psy\VersionUpdater\Checker;
+namespace Psy\VersionUpdater;
 
 use Psy\Shell;
-use Psy\VersionUpdater\Checker;
 
 class GitHubChecker implements Checker
 {
